+++ legacy-zebra/src/learn.rs
@@ -17,7 +17,7 @@ use crate::src::display::TO_SQUARE;
 */
 
 pub unsafe fn init_learn(file_name: *const i8, is_binary: i32, g_state: &mut FullState) {
-    init_osf(0 as i32, g_state);
+    init_osf(0, g_state);
     if is_binary != 0 {
         read_binary_database(file_name, &mut g_state.g_book);
     } else { read_text_database(file_name, &mut g_state.g_book); }
@@ -51,7 +51,7 @@ pub unsafe fn learn_game(game_length: i32,
                                     private_game: i32,
                                     save_database: i32, echo:i32, g_state: &mut FullState) {
     (g_state.g_timer).clear_panic_abort();
-    (g_state.g_timer).toggle_abort_check(0 as i32);
+    (g_state.g_timer).toggle_abort_check(0);
     let full_solve = (g_state.end_g).get_earliest_full_solve();
     let wld_solve = (g_state.end_g).get_earliest_wld_solve();
     let mut dummy: i32 = 0;
@@ -60,31 +60,31 @@ pub unsafe fn learn_game(game_length: i32,
     let mut i = 0;
     while i < game_length {
         generate_all(side_to_move, &mut (g_state.moves_state), &(g_state.search_state), &(g_state.board_state).board);
-        if (g_state.moves_state).move_count[(g_state.moves_state).disks_played as usize] == 0 as i32 {
-            side_to_move = 0 as i32 + 2 as i32 - side_to_move;
+        if (g_state.moves_state).move_count[(g_state.moves_state).disks_played as usize] == 0 {
+            side_to_move = 0 + 2 - side_to_move;
             generate_all(side_to_move, &mut (g_state.moves_state), &(g_state.search_state), &(g_state.board_state).board);
         }
         if (g_state.learn_state).game_move[i as usize] as i32 == -1 {
             fatal_error!("Cannot learn game. Missing move no. {}", i);
         }
         make_move(side_to_move, (g_state.learn_state).game_move[i as usize] as i8,
-                  1 as i32, &mut (g_state.moves_state), &mut (g_state.board_state), &mut (g_state.hash_state), &mut (g_state.flip_stack_));
-        if side_to_move == 2 as i32 {
+                  1, &mut (g_state.moves_state), &mut (g_state.board_state), &mut (g_state.hash_state), &mut (g_state.flip_stack_));
+        if side_to_move == 2 {
             (g_state.learn_state).game_move[i as usize] =
                 -((g_state.learn_state).game_move[i as usize] as i32) as i16
         }
-        side_to_move = 0 as i32 + 2 as i32 - side_to_move;
+        side_to_move = 0 + 2 - side_to_move;
         i += 1
     }
    (g_state.g_book).set_search_depth((g_state.learn_state).learn_depth);
     add_new_game(game_length, None, (g_state.learn_state).cutoff_empty,
-                 full_solve, wld_solve, 1 as i32, private_game, echo, g_state);
+                 full_solve, wld_solve, 1, private_game, echo, g_state);
     if save_database != 0 {
         if g_state.learn_state.binary_database != 0 {
             write_binary_database(g_state.learn_state.database_name.as_mut_ptr(), &mut g_state.g_book);
         } else { write_text_database(g_state.learn_state.database_name.as_mut_ptr(), &mut g_state.g_book); }
     }
-    (g_state.g_timer).toggle_abort_check(1 as i32);
+    (g_state.g_timer).toggle_abort_check(1);
 }
 /*
   FULL_LEARN_PUBLIC_GAME
@@ -116,7 +116,7 @@ pub unsafe fn full_learn_public_game(moves: &[i32],
         fclose(stream);
     }
     ( g_state.g_timer).clear_panic_abort();
-    ( g_state.g_timer).toggle_abort_check(0 as i32);
+    ( g_state.g_timer).toggle_abort_check(0);
     /* Copy the move list from the caller as it is modified below. */
     let mut i = 0;
     while i < length {
@@ -130,26 +130,26 @@ pub unsafe fn full_learn_public_game(moves: &[i32],
     let mut i = 0;
     while i < length {
         generate_all(side_to_move, &mut ( g_state.moves_state), &( g_state.search_state), &( g_state.board_state).board);
-        if ( g_state.moves_state).move_count[( g_state.moves_state).disks_played as usize] == 0 as i32 {
-            side_to_move = 0 as i32 + 2 as i32 - side_to_move;
+        if ( g_state.moves_state).move_count[( g_state.moves_state).disks_played as usize] == 0 {
+            side_to_move = 0 + 2 - side_to_move;
             generate_all(side_to_move, &mut ( g_state.moves_state), &( g_state.search_state), &( g_state.board_state).board);
         }
         make_move(side_to_move, ( g_state.learn_state).game_move[i as usize] as i8,
-                  1 as i32, &mut ( g_state.moves_state), &mut ( g_state.board_state), &mut ( g_state.hash_state), &mut ( g_state.flip_stack_));
-        if side_to_move == 2 as i32 {
+                  1, &mut ( g_state.moves_state), &mut ( g_state.board_state), &mut ( g_state.hash_state), &mut ( g_state.flip_stack_));
+        if side_to_move == 2 {
             ( g_state.learn_state).game_move[i as usize] =
                 -(( g_state.learn_state).game_move[i as usize] as i32) as i16
         }
-        side_to_move = 0 as i32 + 2 as i32 - side_to_move;
+        side_to_move = 0 + 2 - side_to_move;
         i += 1
     }
     /* Let the learning sub-routine in osfbook update the opening
        book and the dump it to file. */
    ( g_state.g_book).set_search_depth(deviation_depth);
     add_new_game(length as _, None, cutoff, exact, wld,
-                 1 as i32, 0 as i32, echo, g_state);
+                 1, 0, echo, g_state);
     if g_state.learn_state.binary_database != 0 {
         write_binary_database(g_state.learn_state.database_name.as_mut_ptr(), &mut g_state.g_book);
     } else { write_text_database(g_state.learn_state.database_name.as_mut_ptr(), &mut g_state.g_book); }
-    ( g_state.g_timer).toggle_abort_check(1 as i32);
+    ( g_state.g_timer).toggle_abort_check(1);
 }
